requirements.txt
fastapi
uvicorn
aiomysql
mysql-connector-python
redis
sqlalchemy
numpy
scikit-learn
sentence-transformers
docker
beautifulsoup4
requests
faiss-cpu
greenlet
cryptography